Write-up 2/requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html

absl-py==0.12.0
albumentations==0.5.2
antlr4-python3-runtime==4.8
appdirs==1.4.4
argon2-cffi==20.1.0
asn1crypto==0.24.0
astunparse==1.6.3
async-generator==1.10
attrs==17.4.0
audioread==2.1.9
Automat==0.6.0
backcall==0.2.0
bleach==3.3.0
blinker==1.4
cached-property==1.5.2
cachetools==4.2.2
catalyst==21.4.2
certifi==2018.1.18
cffi==1.14.5
chardet==3.0.4
classification-models-3D==1.0.2
click==6.7
cloud-tpu-client==0.10
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.1.4
cycler==0.10.0
Cython==0.29.23
dataclasses==0.8
datasets==1.6.2
decorator==4.4.2
defusedxml==0.7.1
dill==0.3.3
distro-info===0.18ubuntu0.18.04.1
efficientnet==1.0.0
efficientnet-3D==1.0.1
efficientnet-pytorch==0.6.3
entrypoints==0.3
fairseq==0.10.2
filelock==3.0.12
flatbuffers==1.12
fsspec==2021.5.0
gast==0.3.3
gdown==3.12.2
gensim==4.0.1
google-api-core==1.26.3
google-api-python-client==1.8.0
google-auth==1.30.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
googleapis-common-protos==1.53.0
grpcio==1.32.0
h5py==2.10.0
httplib2==0.19.1
huggingface-hub==0.0.8
hydra-core==1.0.6
hyperlink==17.3.1
idna==2.6
image-classifiers==1.0.0
imageio==2.9.0
imgaug==0.4.0
importlib-metadata==4.0.1
importlib-resources==5.1.2
incremental==16.10.1
ipykernel==5.5.3
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.10
jiwer==2.2.0
joblib==1.0.1
jsonpatch==1.16
jsonpointer==1.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kaggle==1.5.12
Keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.3.1
language-selector==0.1
librosa==0.8.0
llvmlite==0.36.0
Markdown==3.3.4
MarkupSafe==1.0
matplotlib==3.3.4
mistune==0.8.4
ml-metrics==0.1.4
multiprocess==0.70.11.1
munch==2.5.0
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
netifaces==0.10.4
networkx==2.5.1
nltk==3.6.2
notebook==6.3.0
numba==0.53.1
numexpr==2.7.3
numpy==1.19.5
oauth2client==4.1.3
oauthlib==3.1.0
omegaconf==2.0.6
opencv-python==4.5.1.48
opencv-python-headless==4.5.1.48
opt-einsum==3.3.0
packaging==20.9
PAM==0.4.2
pandas==1.1.5
pandocfilters==1.4.3
parso==0.8.2
pexpect==4.2.1
pickleshare==0.7.4
Pillow==8.2.0
pooch==1.3.0
portalocker==2.0.0
pretrainedmodels==0.7.4
prometheus-client==0.10.1
prompt-toolkit==3.0.18
protobuf==3.15.8
ptyprocess==0.7.0
pyarrow==4.0.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.20
pycrypto==2.6.1
Pygments==2.8.1
pygobject==3.26.1
PyJWT==1.5.3
pyOpenSSL==17.5.0
pyparsing==2.4.7
pyserial==3.4
PySocks==1.7.1
PySoundFile==0.9.0.post1
python-dateutil==2.8.1
python-debian==0.1.32
python-Levenshtein==0.12.2
python-slugify==5.0.0
python-speech-features==0.6
pytz==2021.1
PyWavelets==1.1.1
pyxdg==0.25
PyYAML==5.4.1
pyzmq==22.0.3
qtconsole==5.0.3
QtPy==1.9.0
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
requests-unixsocket==0.1.5
resampy==0.2.2
rouge==1.0.0
rsa==4.7.2
sacrebleu==1.5.1
sacremoses==0.0.45
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
seaborn==0.11.1
SecretStorage==2.3.1
segmentation-models==1.0.1
segmentation-models-pytorch==0.1.3
Send2Trash==1.5.0
sentencepiece==0.1.95
service-identity==16.0.0
Shapely==1.7.1
simplegeneric==0.8.1
six==1.15.0
smart-open==5.0.0
sos==4.1
sounddevice==0.4.1
SoundFile==0.10.3.post1
ssh-import-id==5.7
systemd-python==234
tables==3.6.1
tensorboard==2.5.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.0
tensorboardX==2.2
tensorflow==2.4.1
tensorflow-addons==0.12.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
terminado==0.9.4
testpath==0.4.4
text-unidecode==1.3
threadpoolctl==2.1.0
tifffile==2020.9.3
timm==0.3.2
tokenizers==0.10.2
torch==1.8.1+cu111
torchaudio==0.8.1
torchvision==0.9.1+cu111
tornado==6.1
tqdm==4.49.0
traitlets==4.3.2
transformers==4.5.1
Twisted==17.9.0
typeguard==2.12.0
typing-extensions==3.7.4.3
ufw==0.36
unattended-upgrades==0.1
uritemplate==3.0.1
urllib3==1.22
validators==0.18.2
vecstack==0.4.0
vit-keras==0.0.14
WALinuxAgent==2.2.45
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xgboost==1.4.2
xxhash==2.0.2
zipp==3.4.1
zope.interface==4.3.2
